Extracted smart_screen functionality to own file
