+++ examples/smart_screen.rs
@@ -44,4 +44,3 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
 
     Ok(())
 }
-
