+++ src/client.rs
@@ -1,30 +1,12 @@
+use crate::{RequestError, ResponseError};
 use http::header::CONTENT_TYPE;
 use http::{Request, Response, Uri};
-use serde_json::Value;
 use url::form_urlencoded;
 
-#[derive(thiserror::Error, Debug)]
-#[allow(clippy::upper_case_acronyms)]
-pub enum RequestError {
-    #[error("The provided Uri was invalid")]
-    InvalidUri,
-
-    #[error(transparent)]
-    HTTP(#[from] http::Error),
-}
-
-#[derive(thiserror::Error, Debug)]
-#[allow(clippy::upper_case_acronyms)]
-pub enum ResponseError {
-    #[error("The response status was invalid: {0}")]
-    HttpStatus(http::StatusCode),
-    #[error(transparent)]
-    Deserialize(#[from] serde_json::Error),
-    #[error("Missing field {0} in response")]
-    MissingField(String),
-    #[error("Invalid field {0} in response")]
-    InvalidField(String),
-}
+pub use crate::smart_screen::smart_screen_request;
+pub use crate::smart_screen::smart_screen_result;
+pub use crate::smart_screen::ApiSmartScreenResponse;
+pub use crate::smart_screen::ApiSmartScreenResponseSentiment;
 
 #[derive(Clone, Copy)]
 pub enum Region {
@@ -44,7 +26,7 @@ pub enum Method {
     SmartScreen(String, bool, bool),
 }
 
-fn api_url_by_region(region: Region) -> String {
+pub(crate) fn api_url_by_region(region: Region) -> String {
     match region {
         Region::Us => "https://api1.webpurify.com/services/rest/",
         Region::Europe => "https://api1-eu.webpurify.com/services/rest/",
@@ -58,7 +40,7 @@ fn api_url_by_region(region: Region) -> String {
 ///     Check - returns 1 if profanity is found, otherwise 0
 ///     Replace - returns 1 if profanity if found and replaces
 pub fn query_string(api_key: &str, text: &str, method: Method) -> String {
-    let method_str = match method.clone() {
+    let method_str = match method {
         Method::Check => "webpurify.live.check".to_string(),
         Method::Replace(_) => "webpurify.live.replace".to_string(),
         // TODO(mathias): Change to actually convey meaning about the args (text, sentiment, topics)
@@ -80,7 +62,7 @@ pub fn query_string(api_key: &str, text: &str, method: Method) -> String {
         qs.append_pair("replacesymbol", &replace_with);
     }
 
-    if let Method::SmartScreen(replace_with, _sentiment, _topics) = method.clone() {
+    if let Method::SmartScreen(replace_with, _sentiment, _topics) = method {
         qs.append_pair("replacesymbol", &replace_with);
         qs.append_pair("sentiment", "true");
         qs.append_pair("topics", "true");
@@ -89,11 +71,11 @@ pub fn query_string(api_key: &str, text: &str, method: Method) -> String {
     qs.finish()
 }
 
-pub(crate) fn into_uri<U: TryInto<Uri>>(uri: U) -> Result<Uri, RequestError> {
+pub fn into_uri<U: TryInto<Uri>>(uri: U) -> Result<Uri, RequestError> {
     uri.try_into().map_err(|_err| RequestError::InvalidUri)
 }
 
-fn request_builder(api_uri: String) -> Result<Request<Vec<u8>>, RequestError> {
+pub(crate) fn request_builder(api_uri: String) -> Result<Request<Vec<u8>>, RequestError> {
     let request_builder = Request::builder()
         .method("POST")
         .uri(into_uri(api_uri)?)
@@ -168,20 +150,6 @@ pub fn profanity_replace_request(
     Ok(req)
 }
 
-pub fn smart_screen_request(
-    api_key: &str,
-    region: Region,
-    text: &str,
-    replace_text: &str,
-) -> Result<Request<Vec<u8>>, RequestError> {
-    let qs = query_string(api_key, text, Method::SmartScreen(replace_text.to_string(), true, true));
-    let api_uri = format!("{}?{}", api_url_by_region(region), qs);
-
-    let req = request_builder(api_uri)?;
-    Ok(req)
-}
-
-
 #[derive(serde::Deserialize)]
 struct ApiResponse {
     rsp: ApiResponseRsp,
@@ -197,63 +165,6 @@ fn parse_response<T>(response: Response<T>) -> Result<ApiResponse, ResponseError
 where
     T: AsRef<[u8]>,
 {
-    if !response.status().is_success() {
-        return Err(ResponseError::HttpStatus(response.status()));
-    }
-
-    let body = response.body();
-    Ok(serde_json::from_slice(body.as_ref())?)
-}
-
-#[derive(Debug, serde::Deserialize)]
-#[cfg_attr(test, derive(PartialEq, Eq))]
-pub struct ApiSmartScreenResponseSentiment {
-    text: String,
-    polarity: String,
-}
-
-#[derive(Debug, serde::Deserialize)]
-#[cfg_attr(test, derive(PartialEq, Eq))]
-pub struct ApiSmartScreenResponse {
-    #[serde(deserialize_with="de_bool")]
-    pub bigotry: bool,
-    #[serde(deserialize_with="de_bool")]
-    pub personal_attack: bool,
-    #[serde(deserialize_with="de_bool")]
-    pub sexual_advances: bool,
-    #[serde(deserialize_with="de_bool")]
-    pub criminal_activity: bool,
-    #[serde(deserialize_with="de_bool")]
-    pub external_contact: bool,
-    #[serde(deserialize_with="de_bool")]
-    pub profanity: bool,
-    pub profanity_found: Option<Vec<String>>,
-    pub replace_text: Option<String>,
-
-    /// Only active if "topics=true" is passed to WebPurify in the query string
-    pub topics: Option<Vec<String>>,
-
-    /// Only active if "sentiment=true" is passed to WebPurify in the query string
-    pub overall_sentiment: Option<String>,
-    /// Only active if "sentiment=true" is passed to WebPurify in the query string
-    pub sentiment: Option<Vec<ApiSmartScreenResponseSentiment>>,
-}
-
-fn de_bool<'de, D>(deserializer: D) -> Result<bool, D::Error>
-    where
-        D: serde::de::Deserializer<'de>,
-{
-    let s: String = serde::de::Deserialize::deserialize(deserializer)?;
-
-    match s.as_str() {
-        "true" => Ok(true),
-        "false" => Ok(false),
-        _ => Err(serde::de::Error::unknown_variant(&s, &["true", "false"])),
-    }
-}
-
-fn parse_smart_screen_response<T>(response: Response<T>) -> Result<ApiSmartScreenResponse, ResponseError>
-where T: AsRef<[u8]>, {
     if !response.status().is_success() {
         return Err(ResponseError::HttpStatus(response.status()));
     }
@@ -303,14 +214,6 @@ where
         Some(text) => Ok(text),
         None => Err(ResponseError::MissingField("text".to_owned())),
     }
-}
-
-pub fn smart_screen_result<T>(response: Response<T>) -> Result<ApiSmartScreenResponse, ResponseError>
-where
-    T: AsRef<[u8]>,
-{
-    let res = parse_smart_screen_response(response)?;
-    Ok(res)
 }
 
 #[cfg(test)]
@@ -401,50 +304,6 @@ mod test {
         let result = client::profanity_replace_result(response)?;
 
         assert_eq!(result, "foo".to_owned());
-        Ok(())
-    }
-
-    #[test]
-    fn smart_screen_request() -> Result<(), Box<dyn Error>> {
-        let region = client::Region::Europe;
-        let req = client::smart_screen_request("abcd", region, "hi there", "*")?;
-        assert!(uri_contains(&req, "method=webpurify.live.smartscreen"));
-        assert!(uri_contains(&req, "replacesymbol=*"));
-        assert!(uri_contains(&req, "text=hi+there"));
-
-        Ok(())
-    }
-
-    #[test]
-    fn smart_screen_result() -> Result<(), Box<dyn Error>> {
-        let body = b"{\"language\":\"en\",\"bigotry\":\"false\",\"personal_attack\":\"false\",\"sexual_advances\":\"false\",\
-                                  \"criminal_activity\":\"false\",\"external_contact\":\"false\",\"mental_health\":\"false\",\"profanity\":\"true\",\
-                                  \"profanity_found\":[\"hell\"],\"replace_text\":\"To **** and back\",\"overall_sentmient\":\"negative\",\
-                                  \"sentiment\":[{\"text\":\"to hell and back\",\"polarity\":\"negative\"}]}";
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .body((*body).into_iter().collect::<Vec<_>>())?;
-        let result = client::smart_screen_result(response)?;
-
-        assert_eq!(
-            result,
-            client::ApiSmartScreenResponse {
-                bigotry: false,
-                personal_attack: false,
-                sexual_advances: false,
-                criminal_activity: false,
-                external_contact: false,
-                profanity: true,
-                profanity_found: Some(vec!["hell".to_owned()]),
-                replace_text: Some("To **** and back".to_owned()),
-                topics: None,
-                overall_sentiment: None,
-                sentiment: Some(vec![client::ApiSmartScreenResponseSentiment {
-                    text: "to hell and back".to_owned(),
-                    polarity: "negative".to_owned()
-                }])
-            }
-        );
         Ok(())
     }
 }
